+++ consignments/controllers/HomeController.php
@@ -7,15 +7,9 @@ use Modules\Base\Controller\PageController;
 
 final class HomeController extends PageController
 {
-    public function __construct()
-    {
-        parent::__construct();
-        $this->layout = dirname(__DIR__, 2) . '/_base/views/layouts/cis-template-bare.php';
-    }
-
     public function index(): string
     {
-        return $this->renderView('home/index.php', [
+        return $this->view(dirname(__DIR__) . '/views/home/index.php', [
             'title' => 'Transfers Home',
         ]);
     }